Modified bugs

+++ trading_sales_view.php
@@ -16,15 +16,33 @@ if (isset($_POST['ts_update'])){
         echo "<script>alert('Please Fill-up Contact Number Properly');</script>";
     } else {
         $arr = json_decode($_POST['ts_item_array']);
-        $db_obj_updateTradingSales = new Update_Trading_Sales($trading_sales_number);
-        $db_obj_updateTradingSales->returnStockToWarehouse();
-        $db_obj_updateTradingSales->deleteTradingSalesItem($arr);
-        $db_obj_updateTradingSales->updateTradingSalesItems($arr);
-        $db_obj_updateTradingSales->updateTradingSalesInformation($_POST['ts_number'], $_POST['ts_clientName'], $_POST['ts_representative'], $_POST['ts_contact'],
-        $_POST['ts_address'],$_POST['ts_date'],$_POST['ts_tin'],$_POST['ts_cod']);
-        $db_obj_updateTradingSales->withdrawStocksFromWarehouse($_POST['ts_number'],$_POST['ts_date'],$_POST['ts_clientName'],$arr);
+        $db_obj1 = new Create_New_Trading_Sales;
+        $all_products_available = true;
+        //Check if stock is enought for all items
+        foreach($arr as $items[]) {
+            foreach($items as $item) {
+                // $db_obj1->addTradingSalesItems($item[0],$item[1],$item[2],$item[3]);
+                $product_stock = $db_obj1->checkWarehouseStock($item[0]);
+                if (($product_stock - $item[2]) < 0) {
+                    $all_products_available = false;
+                    echo "<script>alert('Not enough stock');</script>";
+                    break;
+                }
+            }
+        }
+
+        if ($all_products_available) {
+            $db_obj_updateTradingSales = new Update_Trading_Sales($trading_sales_number);
+            $db_obj_updateTradingSales->returnStockToWarehouse();
+            $db_obj_updateTradingSales->deleteTradingSalesItem($arr);
+            $db_obj_updateTradingSales->updateTradingSalesItems($arr);
+            $db_obj_updateTradingSales->updateTradingSalesInformation($_POST['ts_number'], $_POST['ts_clientName'], $_POST['ts_representative'], $_POST['ts_contact'],
+            $_POST['ts_address'],$_POST['ts_date'],$_POST['ts_tin'],$_POST['ts_cod']);
+            $db_obj_updateTradingSales->withdrawStocksFromWarehouse($_POST['ts_number'],$_POST['ts_date'],$_POST['ts_clientName'],$arr);
+            
+            // header('Reload: 0');
+        }
         
-        header('Reload: 0');
     }
     
 }
@@ -129,6 +147,7 @@ if (isset($_POST['ts_delete'])) {
     <!-- Bootstrap CSS -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
     <link rel="stylesheet" href="css/main.css">
+    <script src="js/main.js"></script>
     <title>View Trading Sales</title>
   </head>
   <body>
@@ -149,7 +168,7 @@ if (isset($_POST['ts_delete'])) {
                 </div>
                 <div class="form-group col-md-4">
                     <label for="ts_date">Date(mm/dd/yyyy) </label>
-                    <input class="form-control" id="ts_date" name="ts_date" value="<?php echo $ts_information['trading_sales_date'];?>" readonly/>
+                    <input class="form-control" id="ts_date" name="ts_date" value="<?php echo date("Y-m-d", strtotime($ts_information['trading_sales_date'])) ;?>" readonly/>
                 </div>
                 <div class="form-group col-md-4">
                     <label for="ts_representative">Representative</label>
@@ -172,11 +191,11 @@ if (isset($_POST['ts_delete'])) {
                     <select class="form-control" id="ts_cod" name="ts_cod">
                         <label for="jo_cod">Terms of Payment</label>
                         <option value="COD" <?php if($ts_information['terms_of_payment'] == "COD") echo 'selected="selected"';?>>COD</option>
-                        <option value="30" <?php if($ts_information['terms_of_payment'] == "30") echo 'selected="selected"';?>>30</option>
-                        <option value="60" <?php if($ts_information['terms_of_payment'] == "60") echo 'selected="selected"';?>>60</option>
-                        <option value="90" <?php if($ts_information['terms_of_payment'] == "90") echo 'selected="selected"';?>>90</option>
-                        <option value="150" <?php if($ts_information['terms_of_payment'] == "150") echo 'selected="selected"';?>>150</option>
-                        <option value="180" <?php if($ts_information['terms_of_payment'] == "180") echo 'selected="selected"';?>>180</option>
+                        <option value="30" <?php if($ts_information['terms_of_payment'] == "30") echo 'selected="selected"';?>>30 Days</option>
+                        <option value="60" <?php if($ts_information['terms_of_payment'] == "60") echo 'selected="selected"';?>>60 Days</option>
+                        <option value="90" <?php if($ts_information['terms_of_payment'] == "90") echo 'selected="selected"';?>>90 Days</option>
+                        <option value="150" <?php if($ts_information['terms_of_payment'] == "150") echo 'selected="selected"';?>>150 Days</option>
+                        <option value="180" <?php if($ts_information['terms_of_payment'] == "180") echo 'selected="selected"';?>>180 Days</option>
                     </select>
                 </div> 
             </div>
@@ -232,7 +251,13 @@ if (isset($_POST['ts_delete'])) {
                 </div>
                 <div class="form-group col-md-2 col-sm-6">
                     <label for="ts_unit">Unit</label>
-                    <input class="form-control" id="ts_unit" name="ts_unit" />
+                    <select class="form-control" id="ts_unit" name="ts_unit">
+                        <option>SQM</option>
+                        <option>PC</option>
+                        <option>BAGS</option>
+                        <option>KG</option>
+                        <option>BOX</option>
+                    </select>
                 </div>
                 <div class="form-group col-md-2 col-sm-6">
                     <label for="ts_quantity">Qty.</label>
@@ -266,9 +291,9 @@ if (isset($_POST['ts_delete'])) {
                         <tr>
                             <td><?php echo $tsb_order_item['description'] ?></td>
                             <td><?php echo $tsb_order_item['unit'] ?></td>
-                            <td><?php echo $tsb_order_item['quantity'] ?></td>
-                            <td><?php echo $tsb_order_item['unit_price'] ?></td>
-                            <td><?php echo $tsb_order_item['quantity']*$tsb_order_item['unit_price'] ?></td>
+                            <td><?php echo number_format($tsb_order_item['quantity'],2) ?></td>
+                            <td><?php echo number_format($tsb_order_item['unit_price'],2) ?></td>
+                            <td><?php echo number_format($tsb_order_item['quantity']*$tsb_order_item['unit_price'],2) ?></td>
                             <td><button type='button' class='btn btn-outline-danger btn-sm' onClick='deleteRow(this)'>Delete</button></td>
                         </tr>
